Add recordId to chat navigation and message sending
The code has been updated to include a recordId in the chat navigation parameters within the ToolDetailsScreen.tsx file. Additionally, a new function, postNewMessage, was added in the ChatScreen.tsx file to handle the posting of new messages using either chatId or recordId. If no ID is present, an error alert will notify the user. The sendMessage trigger is now tied to the postNewMessage function.